requirements.txt
pycryptodome>=3.15.0
pyotp>=2.8.0
qrcode[pil]>=7.4.2
Pillow>=9.0.0